manage quiz and manage response

+++ database/migrations/2025_01_18_161612_create_quizzes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->boolean('is_public')->default(true);
             $table->timestamp('start_time')->nullable();
             $table->timestamp('end_time')->nullable();
-            $table->integer('timer')->nullable();
+            $table->float('timer')->nullable();
             $table->integer('total_question');
             $table->timestamps();
         });
